feat: Implement Financial Hub refactoring with three-principle methodology
🏗️ PRINCIPLE 1: Component Breakdown
- Created 5 feature components for Financial hub modularization
- FinancialStatsOverview: Header and financial KPI metrics display
- FinancialQuickActions: Quick action cards for financial operations
- FinancialModuleCards: Main module management with statistics
- FinancialHealthOverview: Financial health metrics and allocation breakdown
- FinancialDataInsights: SSR-safe data fetching and state management

🎣 PRINCIPLE 2: Logic Extraction
- Separated data fetching logic into FinancialDataInsights component
- SSR-safe localStorage access with proper browser guards
- Comprehensive error handling and loading states
- Clean separation of presentation from business logic

🧩 PRINCIPLE 3: UI Component Foundation
- Established reusable component patterns for financial modules
- Professional design consistency maintained throughout
- TypeScript interfaces for type safety and maintainability
- Modular export structure for easy component reuse

⚡ PERFORMANCE & DESIGN
- Reduced Financial hub from 373 lines to modular components
- Maintained RK Institute professional design standards
- Preserved all original functionality and user experience
- Enhanced code maintainability and reusability

✅ STRATEGIC REFACTORING SUCCESS
- Applied proven People hub methodology to Financial hub
- Followed incremental approach with immediate testing
- Maintained design consistency as core priority
- Ready for Vercel deployment validation

This demonstrates the scalability of our three-principle methodology
across different hub types while maintaining professional standards.
